NEW: Need Credentials to enter page and refactor
